+++ include/chess-component/knight.h
@@ -7,7 +7,7 @@ class Knight final : public Piece {
     public:
         Knight(bool white);
 
-        void move(uint8_t side) override;
+        void move(int side) override;
 };
 
 #endif // _KGNIGHT_H_
